Some Change For Constants

+++ Source/Jobs/Controllers/SaleDeliveryOrderSettingsController.cs
@@ -108,7 +108,7 @@ namespace Jobs.Controllers
 
                     LogActivity.LogActivityDetail(LogVm.Map(new ActiivtyLogViewModel
                     {
-                        DocTypeId = pt.DocTypeId,
+                        DocTypeId = (int)pt.DocTypeId,
                         DocId = pt.SaleDeliveryOrderSettingsId,
                         ActivityType = (int)ActivityTypeContants.SettingsAdded,
                     }));
@@ -162,7 +162,7 @@ namespace Jobs.Controllers
 
                     LogActivity.LogActivityDetail(LogVm.Map(new ActiivtyLogViewModel
                     {
-                        DocTypeId = temp.DocTypeId,
+                        DocTypeId = (int)temp.DocTypeId,
                         DocId = temp.SaleDeliveryOrderSettingsId,
                         ActivityType = (int)ActivityTypeContants.SettingsModified,
                         xEModifications = Modifications,
@@ -243,7 +243,7 @@ namespace Jobs.Controllers
 
                     LogActivity.LogActivityDetail(LogVm.Map(new ActiivtyLogViewModel
                     {
-                        DocTypeId = pt.DocTypeId,
+                        DocTypeId = (int)pt.DocTypeId,
                         DocId = pt.SaleDeliveryOrderSettingsId,
                         ActivityType = (int)ActivityTypeContants.SettingsAdded,
                     }));
@@ -297,7 +297,7 @@ namespace Jobs.Controllers
 
                     LogActivity.LogActivityDetail(LogVm.Map(new ActiivtyLogViewModel
                     {
-                        DocTypeId = temp.DocTypeId,
+                        DocTypeId = (int)temp.DocTypeId,
                         DocId = temp.SaleDeliveryOrderSettingsId,
                         ActivityType = (int)ActivityTypeContants.SettingsModified,
                         xEModifications = Modifications,
