+++ Source/Jobs/Controllers/SaleQuotationSettingsController.cs
@@ -121,7 +121,7 @@ namespace Jobs.Controllers
 
                     LogActivity.LogActivityDetail(LogVm.Map(new ActiivtyLogViewModel
                     {
-                        DocTypeId = pt.DocTypeId,
+                        DocTypeId = (int)pt.DocTypeId,
                         DocId = pt.SaleQuotationSettingsId,
                         ActivityType = (int)ActivityTypeContants.SettingsAdded,
                     }));
@@ -205,7 +205,7 @@ namespace Jobs.Controllers
 
                     LogActivity.LogActivityDetail(LogVm.Map(new ActiivtyLogViewModel
                     {
-                        DocTypeId = temp.DocTypeId,
+                        DocTypeId = (int)temp.DocTypeId,
                         DocId = temp.SaleQuotationSettingsId,
                         ActivityType = (int)ActivityTypeContants.SettingsModified,
                         xEModifications = Modifications,
